+++ application/src/scouter/tabs/ScouterAutonomous.tsx
@@ -15,7 +15,7 @@ export const ScouterAutoReef: React.FC = () => {
 
   return (
     <>
-      {ScouterInputs.autoReef.create()}
+      {ScouterInputs.autoReefLevels.create()}
       <button
         className="bg-orange-500 w-20 h-10 text-white py-2 px-4 rounded"
         onClick={() => navigate("../pick")}
