+++ application/src/scouter/tabs/ScouterTeleoperated.tsx
@@ -15,7 +15,7 @@ export const ScouterTeleReef: React.FC = () => {
 
   return (
     <>
-      {ScouterInputs.teleopReef.create()}
+      {ScouterInputs.teleopReefLevels.create()}
       <button
         className="bg-orange-500 w-20 h-10 text-white py-2 px-4 rounded"
         onClick={() => navigate("../pick")}
